+++ drivers/video/tegra/nvmap/nvmap_heap.c
@@ -281,9 +281,9 @@ struct nvmap_heap_block *nvmap_heap_alloc(struct nvmap_heap *h,
 			/* So, page alignment is sufficient check.
 			 */
 			BUG_ON(len & ~(PAGE_MASK));
-			handle->ivm_id = (h->vm_id << NVMAP_IVM_IVMID_SHIFT);
+			handle->ivm_id = ((u64)h->vm_id << NVMAP_IVM_IVMID_SHIFT);
 			handle->ivm_id |= (((offs >> ffs(NVMAP_IVM_ALIGNMENT)) &
-					 ((1 << NVMAP_IVM_OFFSET_WIDTH) - 1)) <<
+					 ((1ULL << NVMAP_IVM_OFFSET_WIDTH) - 1)) <<
 					  NVMAP_IVM_OFFSET_SHIFT);
 			handle->ivm_id |= (len >> PAGE_SHIFT);
 		}
